requirements.txt
requests_async
pycryptodome
QScintilla
websockets
qasync
pyinstaller
pyqt5-tools
PyQt5Designer
cx_freeze